+++ src/global/variables/RandomVariable.java
@@ -1,4 +1,4 @@
-package global.randomframework;
+package global.variables;
 
 public interface RandomVariable {
 }
